fix(learning): make lesson runtime self-contained
