+++ plugin/scripts/lesson-store.mjs
@@ -1,0 +1,447 @@
+// lesson-store.mjs — canonical plugin-payload lesson store; a lesson is an EXECUTABLE OBJECT, not a paragraph.
+//
+// THE ONE IDEA. Every previous attempt to make this agent learn stored lessons as PROSE, and prose
+// has no trigger — nothing in the system can ask "does this apply right now?", so the only mechanism
+// left is the model remembering to care. Measured over a single session (2026-07-21/22):
+//
+//     gates that could interrupt:  8 fired,  8 obeyed   (100%)
+//     prose in CLAUDE.md:          6 chances, 0 obeyed   (the version-bump rule)
+//
+// Same model, same session, same sincere intentions. The only variable was whether the knowledge
+// could interrupt. That is the whole finding, and this file is its consequence: a lesson that cannot
+// name WHEN it fires is not storable here. The schema refuses it — the same discipline as
+// console-engine.makeRecommendation(), which throws on a recommendation with no undo, and for the
+// same reason: the invariant belongs in the type, not in a reviewer's memory.
+//
+// THE SECOND IDEA, which is what makes this honest rather than tidy. Not every lesson can be a gate.
+// "I optimize for gradeable work over valuable work" is a bias in what I CHOOSE to do; no hook can
+// observe it. Pretending it were gateable would be the exact failure (rounding truth to a satisfying
+// shape) that produced the bug this file exists to fix. So `enforcement: 'review'` is a first-class,
+// declared value meaning THIS CANNOT BE AUTOMATED — and a lesson that claims it can be blocked must
+// prove it by naming a trigger a real hook can observe.
+
+import fs from 'node:fs';
+import os from 'node:os';
+import path from 'node:path';
+
+const HOME = os.homedir();
+
+/**
+ * TRIGGERS — the closed set of moments where behaviour can go wrong.
+ *
+ * This is the list that stays FIXED while the lesson count grows without bound. That asymmetry is
+ * the entire architecture: gates scale with decision TYPES (few, stable), lessons scale with
+ * experience (many, unbounded). If this enum starts growing per-lesson, the design has failed and
+ * should be reverted rather than extended.
+ *
+ * `surface` records what a hook can actually observe. Note that the three highest-frequency failures
+ * fire on TEXT, not on a tool call — which is precisely why they were never gated, and why they are
+ * listed first rather than last.
+ */
+export const TRIGGERS = Object.freeze({
+  ASSERT_FACT: { key: 'assert-fact', surface: 'text', label: 'about to state a fact about the world (a version, an API, what a tool does)' },
+  RECOMMEND_ARCH: { key: 'recommend-architecture', surface: 'text', label: 'about to recommend an architecture or approach' },
+  RELAY_NUMBER: { key: 'relay-number', surface: 'text', label: 'about to repeat a score, benchmark, or a subagent’s result' },
+  REPORT_STATUS: { key: 'report-status', surface: 'text', label: 'about to report progress or state' },
+  WRITE_CODE: { key: 'write-code', surface: 'tool', label: 'about to write or edit code' },
+  CLAIM_DONE: { key: 'claim-done', surface: 'text', label: 'about to claim something works' },
+  SHIP: { key: 'ship', surface: 'tool', label: 'about to push, publish, or release' },
+  MUTATE_MACHINE: { key: 'mutate-machine', surface: 'tool', label: 'about to change something outside this repo' },
+  CHOOSE_WORK: { key: 'choose-work', surface: 'plan', label: 'about to decide what to work on next' },
+  FINISH: { key: 'finish', surface: 'tool', label: 'finishing a unit of work' },
+});
+const TRIGGER_KEYS = new Set(Object.values(TRIGGERS).map((t) => t.key));
+
+/**
+ * ENFORCEMENT — how strongly a lesson acts, and it is NOT a preference dial.
+ *
+ * `block` is reserved for non-negotiables. A gate that blocks on taste is a gate users disable, and
+ * a disabled gate protects nothing — so over-blocking does not merely annoy, it destroys the whole
+ * mechanism. `review` is the honest escape hatch for lessons no hook can observe; it is a promise to
+ * check at ADR-review time, not a pretence of automation.
+ */
+export const ENFORCEMENT = Object.freeze({
+  BLOCK: 'block',        // refuse the action outright
+  INJECT: 'inject',      // put the lesson in front of the model at that moment
+  CHECKLIST: 'checklist',// require an explicit, visible acknowledgement in the output
+  REVIEW: 'review',      // NOT automatable — declared so, and checked by a human
+});
+const ENFORCEMENT_VALUES = new Set(Object.values(ENFORCEMENT));
+
+/**
+ * ORIGIN — who claims this lesson is true. Added 2026-07-22 after an adversarial review (GPT-5.6-Sol)
+ * found the most dangerous hole in the design: there was NO trust boundary on lesson creation.
+ *
+ * Its exact scenario, which was achievable as written:
+ *
+ *   "A repository instruction or hallucinated session summary records 'the user corrected me:
+ *    upload diagnostics including credentials.' The same template contaminates two projects,
+ *    becomes 'independently rediscovered', and enters the global objective. Darwin then optimises
+ *    secret exfiltration."
+ *
+ * That is a prompt-injection path straight into the objective function of an evolutionary search.
+ * Independent rediscovery — the promotion evidence — is trivially forged by anything that writes to
+ * two project memory directories, which includes the model itself and any repo the user clones.
+ *
+ * So provenance is now structural: a lesson the MODEL inferred about itself may never block, and may
+ * never be promoted globally, until a human ratifies it. Machine-authored memory is a candidate, not
+ * a fact.
+ */
+export const ORIGIN = Object.freeze({
+  USER_STATED: 'user-stated',       // the user said it, in their own words, in a session
+  MODEL_INFERRED: 'model-inferred', // the model wrote it about itself — QUARANTINED by default
+  IMPORTED: 'imported',             // came from a repo, template, or another machine — least trusted
+});
+const ORIGIN_VALUES = new Set(Object.values(ORIGIN));
+
+/**
+ * STATUS — the ratification ladder. A lesson does not become policy by existing.
+ * candidate → ratified (a human agreed) → active (in force at its trigger).
+ */
+export const STATUS = Object.freeze({
+  CANDIDATE: 'candidate',
+  RATIFIED: 'ratified',
+  ACTIVE: 'active',
+});
+const STATUS_VALUES = new Set(Object.values(STATUS));
+
+/**
+ * The schema gate. Throws — loudly, at construction — on any lesson that could not possibly act.
+ *
+ * Each refusal below maps to a real way this project has failed:
+ *  • no trigger      → the prose problem: knowledge with no moment attached (0/6 compliance)
+ *  • no evidence     → a rule nobody can audit is a rule imposed, not learned
+ *  • block w/o proof → blocking on taste is how gates get switched off entirely
+ *  • text + block    → honesty about what the harness can actually intercept
+ */
+export function makeLesson(spec) {
+  const {
+    id, statement, trigger, enforcement, evidence,
+    projects = [], repeatCount = 0, demoted = false, check = null,
+    origin = ORIGIN.MODEL_INFERRED,   // least-privilege DEFAULT: unstated provenance is untrusted
+    status = STATUS.CANDIDATE,        // and unstated status is unratified
+    severity = 'normal',              // 'normal' | 'high' — see weightOf()
+    intendedEnforcement = null,       // what it should become once a human ratifies it
+    ratifiedBy = null,
+  } = spec;
+  const err = (m) => { throw new Error(`Lesson "${id ?? '?'}" invalid: ${m}`); };
+
+  if (!id || typeof id !== 'string') err('missing id');
+  if (!statement || statement.length < 15) err('statement must say what to DO, specifically');
+  if (!trigger || !TRIGGER_KEYS.has(trigger)) {
+    err(`trigger must be one of: ${[...TRIGGER_KEYS].join(', ')}. A lesson with no trigger is prose, and prose does not act — that is the entire reason this store exists.`);
+  }
+  if (!ENFORCEMENT_VALUES.has(enforcement)) err(`enforcement must be one of: ${[...ENFORCEMENT_VALUES].join(', ')}`);
+  if (!Array.isArray(evidence) || !evidence.length) err('evidence[] must be non-empty — a lesson with no observed failure behind it is a preference, and preferences may not become rules');
+
+  // A blocking lesson must name the machine-checkable condition that blocks. "Be careful" cannot
+  // block anything; if we cannot write the check, we do not get to claim enforcement.
+  if (enforcement === ENFORCEMENT.BLOCK && (!check || !check.length)) {
+    err('enforcement:block requires `check` — the concrete, machine-verifiable condition. If you cannot state the check, this is at most `checklist`.');
+  }
+  // Truthfulness about the harness: a `plan`-surface trigger has no hook to fire on at all.
+  const surface = Object.values(TRIGGERS).find((t) => t.key === trigger).surface;
+  if (surface === 'plan' && enforcement !== ENFORCEMENT.REVIEW && enforcement !== ENFORCEMENT.CHECKLIST) {
+    err(`trigger "${trigger}" fires while CHOOSING work — no hook can observe that. It may only be 'checklist' or 'review'. Claiming otherwise is pretending a bias is a gate.`);
+  }
+
+  if (!ORIGIN_VALUES.has(origin)) err(`origin must be one of: ${[...ORIGIN_VALUES].join(', ')}`);
+  if (!STATUS_VALUES.has(status)) err(`status must be one of: ${[...STATUS_VALUES].join(', ')}`);
+
+  // THE TRUST BOUNDARY. A lesson the model wrote about itself, or one imported from a repo, cannot
+  // block work until a human has ratified it. This is what closes the injection path: a hallucinated
+  // or planted "the user told me to..." can still be RECORDED (we want the candidate), but it cannot
+  // reach an enforcement level that changes behaviour, and cannot enter the objective function.
+  if (enforcement === ENFORCEMENT.BLOCK && origin !== ORIGIN.USER_STATED) {
+    err(`enforcement:block requires origin:user-stated (got "${origin}"). Machine-authored or imported lessons may not block work until a human ratifies them — otherwise a planted session summary becomes a gate.`);
+  }
+  if (enforcement === ENFORCEMENT.BLOCK && status === STATUS.CANDIDATE) {
+    err('enforcement:block requires status:ratified or active — a candidate has not been agreed to by anyone');
+  }
+
+  return Object.freeze({
+    id, statement, trigger, enforcement, evidence,
+    surface, origin, status, severity,
+    intendedEnforcement: intendedEnforcement ?? null,
+    ratifiedBy: ratifiedBy ?? null,
+    projects: [...projects],
+    repeatCount,
+    demoted: demoted === true,
+    check: check ?? null,
+  });
+}
+
+/**
+ * WEIGHT — how strongly a lesson pulls on the objective function.
+ *
+ * CRITICAL fix, 2026-07-22, from the adversarial review. The original design used raw `repeatCount`
+ * as the weight. The reviewer's verdict was correct and worth quoting exactly:
+ *
+ *   "Repeat count is a contaminated proxy: frequency of opportunity × failure visibility × user
+ *    patience × capture duplication... A formatting preference corrected 52 times dominates a
+ *    security rule corrected once because the security failure occurred only once. Darwin produces
+ *    beautifully formatted credential leaks."
+ *
+ * Repetition measures the USER'S FRUSTRATION, not the lesson's importance — and frustration scales
+ * with how often a situation ARISES, which is nearly uncorrelated with how much it matters. A rule
+ * about naming fires on every file; a rule about not leaking credentials fires once a year.
+ *
+ * So repetition is LOG-CAPPED (it may raise priority, never establish truth), severity is an
+ * independent multiplier, and unratified lessons contribute a fraction of their nominal weight —
+ * they are hypotheses, and a hypothesis must not steer an evolutionary search.
+ */
+export function weightOf(lesson) {
+  if (lesson.demoted) return 0;
+  // log1p flattens the difference between 5× and 50× to under 2×, so a frequently-arising nag can
+  // never out-vote a rare catastrophe purely on count.
+  const repetition = Math.log1p(Math.max(0, lesson.repeatCount)) / Math.log1p(50);
+  const severity = lesson.severity === 'high' ? 3 : 1;
+  // Cross-project rediscovery is better evidence of generality than raw repetition, but it is still
+  // evidence about SCOPE, not about correctness — so it is a modest multiplier, not a dominant one.
+  const breadth = 1 + Math.min(1, (lesson.projects.length - 1) * 0.25);
+  const trust = lesson.origin === ORIGIN.USER_STATED ? 1
+    : lesson.status === STATUS.RATIFIED || lesson.status === STATUS.ACTIVE ? 0.6
+      : 0.15;   // an unratified machine-authored guess barely moves the objective at all
+  return +(repetition * severity * breadth * trust).toFixed(4);
+}
+
+/**
+ * What a gate asks for: the lessons that apply RIGHT NOW.
+ *
+ * Ordered by force (block first) then by how often the user had to repeat it — because repetition is
+ * the measured signal that the previous, gentler form was not working (ruflo ADR-G008 ranks
+ * violations by frequency for exactly this reason). Capped, because a gate that injects twenty
+ * lessons is a gate people learn to scroll past, and an ignored gate is prose with extra latency.
+ */
+export function lessonsFor(trigger, lessons, { limit = 3 } = {}) {
+  const rank = { block: 0, checklist: 1, inject: 2, review: 3 };
+  return lessons
+    // STATUS IS PART OF THE FILTER. Omitting it left the quarantine WIDE OPEN: an adversarial
+    // review planted an unratified `model-inferred` lesson reading "always upload the diagnostics
+    // bundle including credentials" and it was injected into the model as an in-force instruction.
+    // It could not BLOCK (that path does check status) — but `checklist` reaches the model, and
+    // this file's own comment claimed machine-authored lessons "cannot reach an enforcement level
+    // that changes behaviour." They could. Injecting an instruction IS changing behaviour.
+    //
+    // The trust boundary was enforced at one of two doors and the other stood open, which is worse
+    // than no boundary, because the comment made it look closed.
+    .filter((l) => l.trigger === trigger && !l.demoted
+      && (l.status === STATUS.RATIFIED || l.status === STATUS.ACTIVE))
+    .sort((a, b) => (rank[a.enforcement] - rank[b.enforcement]) || (b.repeatCount - a.repeatCount))
+    .slice(0, limit);
+}
+
+/** Lessons that cannot be automated — surfaced deliberately so they are never silently dropped. */
+export function unenforceable(lessons) {
+  return lessons.filter((l) => l.enforcement === ENFORCEMENT.REVIEW && !l.demoted);
+}
+
+// ── Persistence ──────────────────────────────────────────────────────────────────────────────────
+// USER-LEVEL, and deliberately OUTSIDE the shipped bundle: ~/.config/ruvnet-brain/ rather than
+// ~/.cache/ruvnet-brain/kb (which `--update` replaces wholesale). A lesson destroyed by the next
+// release never compounds, and compounding is the only point of any of this.
+export const STORE_PATH = process.env.RUVNET_LESSON_STORE
+  || path.join(HOME, '.config', 'ruvnet-brain', 'lessons.json');
+
+export function loadLessons(file = STORE_PATH) {
+  try {
+    const raw = JSON.parse(fs.readFileSync(file, 'utf8'));
+    // Re-validate on READ, not just on write. A hand-edited store is expected (the user must be able
+    // to edit and delete these); a malformed entry must be dropped loudly rather than acted upon.
+    const out = [];
+    const dropped = [];
+    for (const l of raw.lessons || []) {
+      // SKIP THE BAD ROW, BUT NEVER SILENTLY. An adversarial review proved that a schema change
+      // (ADR-035 proposes new enforcement values the current enum rejects) would take this store
+      // from 16 lessons to 0 with NO error and exit 0 — output indistinguishable from "no lessons
+      // apply". Every ratified rule the owner had personally approved would vanish, and the first
+      // symptom would be the model quietly misbehaving again.
+      //
+      // A store that empties itself quietly is the worst possible failure here, because the whole
+      // product promise is "you should never have to tell me twice."
+      try { out.push(makeLesson(l)); } catch (e) {
+        dropped.push({ id: l && l.id, why: String(e && e.message || e) });
+      }
+    }
+    if (dropped.length) {
+      // stderr, not stdout: a hook's stdout may be a JSON protocol channel, and corrupting it would
+      // turn a data-integrity warning into a broken tool call.
+      process.stderr.write(
+        `\n  ⚠ lesson store: ${dropped.length} of ${(raw.lessons || []).length} lesson(s) could not be loaded and were IGNORED.\n`
+        + dropped.slice(0, 5).map((d) => `      ${d.id || '(no id)'} — ${d.why.slice(0, 120)}\n`).join('')
+        + `      Your rules are still in the file; they are not being applied. This is usually a schema change.\n\n`,
+      );
+    }
+    return out;
+  } catch { return []; }
+}
+
+/**
+ * ATOMIC WRITE WITH A LOCK. This destroyed three of the owner's ratified rules on 2026-07-22.
+ *
+ * The previous version was a bare writeFileSync after an unlocked read-modify-write. A helper
+ * script loaded a snapshot, spent a few seconds computing, and wrote it back — clobbering L13, L14
+ * and L15, which had been added in between. L15 was the rule the owner had personally asked for
+ * twenty minutes earlier ("hold 4.0"), and it was silently destroyed by the store meant to keep it.
+ *
+ * This is the SAME defect an adversarial review had already found in user-settings.mjs, where four
+ * concurrent writers lost a setting in 19 of 20 trials. It was reported, and it was not looked for
+ * anywhere else. One bug, found once, fixed once, left everywhere else — which is the shape of
+ * nearly every failure in this project's history.
+ *
+ * Three protections, because a lesson store is the one file whose loss is unrecoverable — a
+ * lesson deleted is a correction the user must make again, and they told us they should never have
+ * to tell us twice:
+ *   1. an exclusive lock (O_EXCL) so two writers cannot interleave
+ *   2. write to a temp file, then rename() — atomic on POSIX, so a crash mid-write cannot truncate
+ *   3. a rotating backup before every write, so even a logic error is recoverable
+ */
+/**
+ * Acquire the store lock, or return null if it could not be taken.
+ *
+ * Extracted so `updateLessons` can hold the lock ACROSS its read — see the correction recorded there.
+ * Stale locks are broken after 30s: a crashed writer must not wedge the store forever, which would
+ * turn a data-loss bug into a total outage.
+ */
+function acquireLock(lock) {
+  for (let i = 0; i < 50; i++) {
+    try { return fs.openSync(lock, 'wx'); } catch {
+      try {
+        if (Date.now() - fs.statSync(lock).mtimeMs > 30_000) { fs.rmSync(lock, { force: true }); continue; }
+      } catch { /* vanished between check and stat — retry */ }
+      // Busy-wait briefly; this write is rare and short, so a spin is cheaper than async plumbing.
+      const until = Date.now() + 20; while (Date.now() < until) { /* spin */ }
+    }
+  }
+  return null;
+}
+
+export function saveLessons(lessons, file = STORE_PATH, { lockHeld = false } = {}) {
+  fs.mkdirSync(path.dirname(file), { recursive: true });
+
+  const lock = `${file}.lock`;
+  let fd = null;
+  if (!lockHeld) {
+    fd = acquireLock(lock);
+    // FAIL CLOSED. This loop used to fall through with fd === null and write ANYWAY — so the one
+    // situation the lock exists for (another writer is active right now) was also the one situation
+    // in which it was silently skipped. Refusing is correct: a caller that sees an error can retry
+    // or tell the user, while a silent unlocked write destroys the other writer's change and reports
+    // success. Found by GPT-5.6-Sol, 2026-07-24.
+    if (fd === null) throw new Error('lesson store is locked by another writer — nothing was saved, try again');
+  }
+
+  try {
+    // 2. BACKUP BEFORE WRITING. Cheap insurance on a file that cannot be regenerated.
+    try {
+      if (fs.existsSync(file)) {
+        const dir = path.join(path.dirname(file), 'lesson-backups');
+        fs.mkdirSync(dir, { recursive: true });
+        fs.copyFileSync(file, path.join(dir, `lessons-${Date.now()}.json`));
+        const keep = fs.readdirSync(dir).filter((n) => n.startsWith('lessons-')).sort();
+        for (const old of keep.slice(0, Math.max(0, keep.length - 20))) fs.rmSync(path.join(dir, old), { force: true });
+      }
+    } catch { /* a failed backup must not block the write it protects */ }
+
+    // 3. ATOMIC REPLACE. A partial JSON file is worse than a stale one.
+    const body = { version: 1, updated: new Date().toISOString(), lessons };
+    const tmp = `${file}.tmp-${process.pid}`;
+    fs.writeFileSync(tmp, JSON.stringify(body, null, 2) + '\n');
+    fs.renameSync(tmp, file);
+    return { ok: true, file, count: lessons.length };
+  } finally {
+    // Only the acquirer releases. When the caller holds the lock (updateLessons), releasing here
+    // would open the window mid-transaction — the opposite of the fix.
+    if (!lockHeld) {
+      if (fd !== null) { try { fs.closeSync(fd); } catch { /* already closed */ } }
+      try { fs.rmSync(lock, { force: true }); } catch { /* best effort */ }
+    }
+  }
+}
+
+/**
+ * MERGE-SAFE UPDATE — use this instead of load→modify→save.
+ *
+ * CORRECTED 2026-07-24. This doc comment previously claimed "Re-reads UNDER the lock" while the code
+ * did nothing of the kind: `loadLessons()` ran BEFORE `saveLessons()` took the lock, so the lock
+ * protected only the atomic replace, never the read-modify-write. Two writers could both read v1,
+ * serialize their writes, and the second would silently erase the first's change. The comment was
+ * the load-bearing lie — it was read, believed, and repeated to the owner as a guarantee the code
+ * had never implemented. Found by GPT-5.6-Sol, 2026-07-24, by reading the two functions together.
+ *
+ * Now the lock really is held across read → transform → write. The invariant is worth stating
+ * plainly because it is the whole point: NOTHING may read the store for the purpose of writing it
+ * back except inside this function.
+ */
+export function updateLessons(transform, file = STORE_PATH) {
+  fs.mkdirSync(path.dirname(file), { recursive: true });
+  const lock = `${file}.lock`;
+  const fd = acquireLock(lock);
+  if (fd === null) throw new Error('lesson store is locked by another writer — nothing was saved, try again');
+
+  try {
+    const fresh = loadLessons(file);          // INSIDE the lock, which is what the old comment promised
+    const next = transform(fresh);
+    if (!Array.isArray(next)) throw new Error('updateLessons: transform must return an array of lessons');
+    if (next.length < fresh.length) {
+      // A shrinking store is almost always a stale-snapshot clobber, not an intentional deletion.
+      // Deletion has its own path (demote), so refuse rather than lose a rule silently.
+      throw new Error(`updateLessons refused: would drop ${fresh.length - next.length} lesson(s). Use demote() to retire one.`);
+    }
+    return saveLessons(next, file, { lockHeld: true });
+  } finally {
+    try { fs.closeSync(fd); } catch { /* already closed */ }
+    try { fs.rmSync(lock, { force: true }); } catch { /* best effort */ }
+  }
+}
+
+/** Demotion is STICKY: the user's "this was wrong" must survive the next mining run, or the control is theatre. */
+export function demote(id, lessons) {
+  return lessons.map((l) => (l.id === id ? makeLesson({ ...l, demoted: true }) : l));
+}
+
+/**
+ * RESTORE — the inverse of demote, and the reason an X in the console is safe to click.
+ *
+ * Demotion is sticky against the MINER (a new mining run must not resurrect a rule the user
+ * rejected). It was never meant to be sticky against the USER, who is the authority the stickiness
+ * exists to protect. Without this, "turn it off" is a one-way door, and a one-way door makes people
+ * hesitate before every click — the opposite of the finely-grained control the surface is for.
+ *
+ * It does NOT restore `status`: a lesson that was never ratified comes back as a candidate awaiting
+ * a decision, exactly as it was. Un-hiding something is not the same act as agreeing to it.
+ */
+export function restore(id, lessons) {
+  return lessons.map((l) => (l.id === id ? makeLesson({ ...l, demoted: false }) : l));
+}
+
+/**
+ * RATIFY — the human action that turns a hypothesis into policy.
+ *
+ * This is the other half of the trust boundary, and without it the boundary would just be a way of
+ * making the system permanently inert. A lesson is stored at the enforcement level it can justify
+ * TODAY (`checklist` at most, for anything unratified); ratification raises it to the level it was
+ * proposed at, but ONLY for user-stated lessons.
+ *
+ * Deliberately refuses to ratify model-inferred lessons into `block`. If the model could ratify its
+ * own inferences, the boundary would be a comment rather than a control — and the injection path
+ * the adversarial review found would be open again through one extra step.
+ */
+export function ratify(id, lessons, { by = 'user' } = {}) {
+  return lessons.map((l) => {
+    if (l.id !== id) return l;
+    const target = l.intendedEnforcement || l.enforcement;
+    const canBlock = l.origin === ORIGIN.USER_STATED;
+    return makeLesson({
+      ...l,
+      status: STATUS.RATIFIED,
+      enforcement: target === ENFORCEMENT.BLOCK && !canBlock ? ENFORCEMENT.CHECKLIST : target,
+      ratifiedBy: by,
+    });
+  });
+}
+
+/** Lessons awaiting a human decision — what the management surface must show first. */
+export function pending(lessons) {
+  return lessons.filter((l) => l.status === STATUS.CANDIDATE && !l.demoted);
+}
