+++ scripts/learning-replay.mjs
@@ -214,7 +214,8 @@ export const LOAD_BEARING = Object.freeze([
   'scripts/ci/learning-replay-recorder.mjs',
   'scripts/ci/learning-replay-codex-adapter.mjs',
   'scripts/lesson-store.mjs',
-  'scripts/lesson-gate.mjs',
+  'plugin/scripts/lesson-store.mjs',
+  'plugin/scripts/lesson-gate.mjs',
   'plugin/scripts/lesson-hooks.sh',
   'plugin/scripts/unprompted-runtime.mjs',
   'plugin/scripts/hook-shim.mjs',
