+++ tests/unit/learning-replay.test.mjs
@@ -569,7 +569,8 @@ describe('--check gates on a STATED SHA, and UNKNOWN is never PASS', () => {
   it('names the files whose change invalidates a recorded result', () => {
     // A currency rule nobody can enumerate is a currency rule nobody can audit.
     expect(LOAD_BEARING).toContain('scripts/learning-replay.mjs');
-    expect(LOAD_BEARING).toContain('scripts/lesson-gate.mjs');
+    expect(LOAD_BEARING).toContain('plugin/scripts/lesson-gate.mjs');
+    expect(LOAD_BEARING).toContain('plugin/scripts/lesson-store.mjs');
     expect(LOAD_BEARING).toContain('plugin/scripts/hook-shim.mjs');
   });
 });
